+++ packages/core/src/components/sections/Hero/Hero.tsx
@@ -1,9 +1,10 @@
+import {
+  Hero as UIHero,
+  HeroHeading as UIHeroHeading,
+  HeroImage as UIHeroImage,
+} from '@faststore/ui'
+
 import { ButtonLink } from 'src/components/ui/Button'
-import UIHero, {
-  HeroContent,
-  HeroImage,
-  HeroLink,
-} from 'src/components/ui/Hero'
 import Icon from 'src/components/ui/Icon'
 import Image from 'src/components/ui/Image/Image'
 
@@ -42,7 +43,7 @@ const Hero = ({
         data-fs-hero-variant={variant}
         data-fs-hero-color-variant={colorVariant}
       >
-        <HeroImage data-fs-hero-image>
+        <UIHeroImage data-fs-hero-image>
           <Image
             preload
             loading="eager"
@@ -53,26 +54,23 @@ const Hero = ({
             height={240}
             sizes="(max-width: 768px) 70vw, 50vw"
           />
-        </HeroImage>
-        <HeroContent data-fs-hero-content aria-labelledby="hero-heading">
+        </UIHeroImage>
+        <UIHeroHeading data-fs-hero-heading aria-labelledby="hero-heading">
           <div data-fs-hero-wrapper className="layout__content">
             <div data-fs-hero-info>
               <h1 id="hero-heading">{title}</h1>
-
               <p data-fs-hero-text-body>{subtitle}</p>
               {!!link && (
-                <HeroLink>
-                  <ButtonLink href={link} inverse={colorVariant === 'main'}>
-                    {linkText} <Icon name="ArrowRight" width={24} height={24} />
-                  </ButtonLink>
-                </HeroLink>
+                <ButtonLink href={link} inverse={colorVariant === 'main'}>
+                  {linkText} <Icon name="ArrowRight" width={24} height={24} />
+                </ButtonLink>
               )}
             </div>
             {variant === 'secondary' && icon && (
               <div data-fs-hero-icon>{icon}</div>
             )}
           </div>
-        </HeroContent>
+        </UIHeroHeading>
       </UIHero>
     </Section>
   )