+++ packages/core/src/pages/index.tsx
@@ -55,7 +55,7 @@ function Page() {
         title="New Offers"
         subtitle="At BaseStore you can shop the best tech of 2022. Enjoy and get 10% off on your first purchase."
         linkText="See all"
-        link="/"
+        link="/technology"
         imageSrc="https://storeframework.vtexassets.com/arquivos/ids/190897/Photo.jpg"
         imageAlt="Quest 2 Controller on a table"
       />